tikv: move temporary export types to package txnkv (#258)
